requirements.txt
Pillow==10.0.0
Pillow==10.0.1
Scrapy==2.10.1
selenium==4.12.0
thefuzz==0.20.0
